float -> double

+++ src/camera.hpp
@@ -4,28 +4,28 @@
 #include "ray.hpp"
 #include "utils.hpp"
 
-// Camera **with a film**
-// TODO: 用fov形式表示film
+// Camera **with a film**lele
+// TODO: 用fov形式表示filmle
 class Camera
 {
-    using Vec3 = Eigen::Vector3f;
+    using Vec3 = Eigen::Vector3d;
 
 protected:
     // Camera intrinsics
-    float _focal = 1.0f;
-    float _aspectRatio = 1024.f / 768.f;
-    float _hfov = degree2radian(100.0f);
+    double _focal = 1.0;
+    double _aspectRatio = 1024.0 / 768.0;
+    double _hfov = degree2radian(100.0);
 
     // Camera Extrinsics
-    Vec3 _pos = Vec3{0.0f, 0.0f, 0.0f};
-    Vec3 _up = Vec3{0.0f, 1.0f, 0.0f};
-    Vec3 _lookAt = Vec3{0.0f, 0.0f, -1.0f};
-    Vec3 _rightHand = Vec3{1.0f, 0.0f, 0.0f};
+    Vec3 _pos = Vec3{0.0, 0.0, 0.0};
+    Vec3 _up = Vec3{0.0, 1.0, 0.0};
+    Vec3 _lookAt = Vec3{0.0, 0.0, -1.0};
+    Vec3 _rightHand = Vec3{1.0, 0.0, 0.0};
 
     // Viewport parameters
     // Should not be changed directly
-    float _width = 0.0f;
-    float _height = 0.0f;
+    double _width = 0.0;
+    double _height = 0.0;
 
     // Film parameters
     int _nHrozPix = 1024;
@@ -33,9 +33,9 @@ protected:
     // Once we modify the camera or the film, the width/height per pixel will change.
     // We set a flag to ensure recompute when use.
     bool _recomputeFilmFlag = true;
-    Vec3 _firstPixelCenter{0.0f, 0.0f, 0.0f};
-    float _wPix; // width per pixel
-    float _hPix; // height per pixel
+    Vec3 _firstPixelCenter{0.0, 0.0, 0.0};
+    double _wPix; // width per pixel
+    double _hPix; // height per pixel
 
 protected:
     void _recomputeFilm()
@@ -43,8 +43,8 @@ protected:
         _wPix = _width / _nHrozPix;
         _hPix = _height / _nVertPix;
         Vec3 filmCenter = _pos + _focal * _lookAt;
-        float offLeft = _width / 2 - 0.5f * _wPix;
-        float offUp = _height / 2 - 0.5f * _hPix;
+        double offLeft = _width / 2 - 0.5 * _wPix;
+        double offUp = _height / 2 - 0.5 * _hPix;
         _firstPixelCenter = filmCenter - _rightHand * offLeft + _up * offUp;
     }
 
@@ -52,27 +52,27 @@ public:
     Camera(){};
 
 public:
-    virtual Ray rayThroughFilm(float row, float col) = 0;
+    virtual Ray rayThroughFilm(double row, double col) = 0;
 
     // parameter setters
 public:
-    inline void setFocal(float focal)
+    inline void setFocal(double focal)
     {
         _focal = focal;
-        _width = 2.0f * _focal * tan(_hfov / 2.0f);
+        _width = 2.0 * _focal * tan(_hfov / 2.0);
         _height = _width / _aspectRatio;
         _recomputeFilmFlag = true;
     }
-    inline void setAspectRatio(float aspectRatio)
+    inline void setAspectRatio(double aspectRatio)
     {
         _aspectRatio = aspectRatio;
         _height = _width / aspectRatio;
         _recomputeFilmFlag = true;
     }
-    inline void setHFOV(float hfov)
+    inline void setHFOV(double hfov)
     {
         _hfov = degree2radian(hfov);
-        _width = 2.0f * _focal * tan(_hfov / 2.0f);
+        _width = 2.0 * _focal * tan(_hfov / 2.0);
         _height = _width / _aspectRatio;
         _recomputeFilmFlag = true;
     }
@@ -110,10 +110,10 @@ public:
 
 class PerspectiveCamera : public Camera
 {
-    using Vec3 = Eigen::Vector3f;
+    using Vec3 = Eigen::Vector3d;
 
 public:
-    Ray rayThroughFilm(float row, float col)
+    Ray rayThroughFilm(double row, double col)
     {
         if (_recomputeFilmFlag)
         {
@@ -127,10 +127,10 @@ public:
 
 class OrthogonalCamera : public Camera
 {
-    using Vec3 = Eigen::Vector3f;
+    using Vec3 = Eigen::Vector3d;
 
 public:
-    Ray rayThroughFilm(float row, float col)
+    Ray rayThroughFilm(double row, double col)
     {
         if (_recomputeFilmFlag)
         {
